[#77]/refactor FolderPathEllipsis 컴포넌트 추출
